Use usize instead of i32

+++ yotsuba/src/utils.rs
@@ -1,4 +1,4 @@
-pub fn pad_sequence(sequence: &Vec<i32>, maxlen: i32, value: i32) -> Vec<i32> {
+pub fn pad_sequence(sequence: &Vec<i32>, maxlen: usize, value: i32) -> Vec<i32> {
     // sequence=[0, 1, 2], maxlen=5, value=10 -> [0, 1, 2, 10, 10]
     let mut ret: Vec<i32> = Vec::new();
     for i in 0..maxlen {
@@ -15,7 +15,7 @@ pub fn pad_sequence(sequence: &Vec<i32>, maxlen: i32, value: i32) -> Vec<i32> {
 }
 
 
-pub fn pad_sequences(sequences: &Vec<Vec<i32>>, maxlen: i32, value: i32) -> Vec<Vec<i32>> {
+pub fn pad_sequences(sequences: &Vec<Vec<i32>>, maxlen: usize, value: i32) -> Vec<Vec<i32>> {
     // sequences=[[0, 1, 2], ...], maxlen=5, value=10 -> [[0, 1, 2, 10, 10], ...]
     let mut ret: Vec<Vec<i32>> = Vec::new();
     for sequence in sequences {
